requirements.txt
pyserial
numpy
invoke
paramiko
fabric3
pyvisa
pyvisa-py
